test_project/pyproject.toml
[project]
name = "test-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "boto3>=1.37.38",
    "litellm>=1.80.0",
    "mlflow>=2.17.2",
    "openai>=2.2.0",
    "pyspark",
    "python-dotenv>=1.0.1",
]
